Create/move files to UI namespace
